operation domain modified for resources
